Rename data files
Try to make names represent data.  So files are named by *observation
type*. For example, if each row in the data file represents a burning
trial, the name should be "burning-trials".  If each row represents a
specimen, the name should begin with "specimens"

Some data files still need to be renamed